#127/Impl OnExitLecture & GetActiveLectureMember
